Move comparator subcomponents to page folder
