+++ src/app/not-found.tsx
@@ -10,10 +10,11 @@ import Image from 'next/image'
 import NextLink from 'next/link'
 import { type FC } from 'react'
 
+import { Link } from '~/components/ChakraNextLink'
+
 import { openGraph } from './shared-metadata'
 
 import mascotIcon from '@app-public/mascot-icon.png'
-import { Link } from '~/components/ChakraNextLink'
 
 export const metadata = {
 	title: 'Not Found',
