+++ src/components/Header.tsx
@@ -14,10 +14,11 @@ import Image from 'next/image'
 import * as React from 'react'
 import { DiGithubBadge } from 'react-icons/di'
 
-import mascotIcon from '@app-public/mascot-icon.png'
 import { REPO_URL } from '~/common'
 import { Link as RouteLink } from '~/components/ChakraNextLink'
 import ToggleColorModeButton from '~/components/ToggleColorModeButton'
+
+import mascotIcon from '@app-public/mascot-icon.png'
 
 const LOGO_SIZES = { base: 8, md: 16 }
